AS7-5677 Split up GlobalOperationHandlers
