+++ controller/src/test/java/org/jboss/as/controller/test/WildcardUnitTestCase.java
@@ -35,6 +35,7 @@ import org.jboss.as.controller.PathAddress;
 import org.jboss.as.controller.PathElement;
 import org.jboss.as.controller.descriptions.DescriptionProvider;
 import org.jboss.as.controller.operations.global.GlobalOperationHandlers;
+import org.jboss.as.controller.operations.global.ReadResourceHandler;
 import org.jboss.as.controller.registry.ImmutableManagementResourceRegistration;
 import org.jboss.as.controller.registry.ManagementResourceRegistration;
 import org.jboss.as.controller.registry.Resource;
@@ -93,7 +94,7 @@ public class WildcardUnitTestCase extends AbstractControllerTestBase {
 
     @Override
     protected void initModel(Resource rootResource, ManagementResourceRegistration root) {
-            root.registerOperationHandler("read-resource", GlobalOperationHandlers.READ_RESOURCE, NULL, true);
+            root.registerOperationHandler("read-resource", ReadResourceHandler.INSTANCE, NULL, true);
             root.registerOperationHandler("describe", new DescribeHandler(), NULL, true);
 
             root.registerOperationHandler("setup", new OperationStepHandler() {
